Move async shadow runner tests
